update to snapshot spandsp-20090130
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@11563 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/spandsp/src/dtmf.c
@@ -22,7 +22,7 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- * $Id: dtmf.c,v 1.47 2009/01/28 03:41:26 steveu Exp $
+ * $Id: dtmf.c,v 1.48 2009/01/29 18:30:14 steveu Exp $
  */
  
 /*! \file dtmf.h */
@@ -485,7 +485,7 @@ size_t dtmf_tx_put(dtmf_tx_state_t *s, const char *digits, int len)
         if ((len = strlen(digits)) == 0)
             return 0;
     }
-    if ((space = queue_free_space(&s->queue.queue)) < len)
+    if ((space = queue_free_space(&s->queue.queue)) < (size_t) len)
         return len - space;
     if (queue_write(&s->queue.queue, (const uint8_t *) digits, len) >= 0)
         return 0;
